# api/openapi.yaml
openapi: 3.1.0
info:
  title: Marketplace Golang API
  description: |
    REST API for the marketplace-golang single-vendor e-commerce backend.
    Phase 1c adds admin image upload with variant generation.
    Phase 2a adds authentication (register, login, verify, reset) and account management endpoints.
  version: 0.4.0
servers:
  - url: http://localhost:8080
    description: Local development
tags:
  - name: auth
    description: Authentication endpoints (register, login, verify, reset).
  - name: account
    description: Authenticated account-management endpoints (/me, sessions).
paths:
  /health:
    get:
      summary: Liveness probe
      operationId: getHealth
      responses:
        "200":
          description: Process alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [ok] }
                required: [status]
  /ready:
    get:
      summary: Readiness probe
      operationId: getReady
      responses:
        "200":
          description: All checks healthy
          content:
            application/json:
              schema: { type: object, additionalProperties: { type: string } }
        "503":
          description: One or more checks failed
          content:
            application/json:
              schema: { type: object, additionalProperties: { type: string } }
  /metrics:
    get:
      summary: Prometheus metrics
      operationId: getMetrics
      responses:
        "200":
          description: Metrics text
          content:
            text/plain:
              schema: { type: string }
  /products:
    get:
      summary: List published products
      operationId: listProducts
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: categoryId
          in: query
          schema: { type: string, format: uuid }
        - name: brand
          in: query
          schema: { type: string }
        - name: minPriceCents
          in: query
          schema: { type: integer, minimum: 0 }
        - name: maxPriceCents
          in: query
          schema: { type: integer, minimum: 0 }
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
  /products/{slug}:
    get:
      summary: Get product by slug
      operationId: getProductBySlug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /search:
    get:
      summary: Free-text product search
      operationId: searchProducts
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Matched products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
        "400":
          description: Invalid query
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /categories:
    get:
      summary: List all categories
      operationId: listCategories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Category" }
  /admin/products:
    post:
      summary: Create a product (admin)
      operationId: adminCreateProduct
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductWriteBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "400":
          description: Invalid input
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /admin/products/{id}:
    put:
      summary: Update a product (admin)
      operationId: adminUpdateProduct
      security: [ { bearerAuth: [] } ]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductWriteBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      summary: Delete a product (admin)
      operationId: adminDeleteProduct
      security: [ { bearerAuth: [] } ]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /admin/categories:
    post:
      summary: Create a category (admin)
      operationId: adminCreateCategory
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryWriteBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
  /admin/products/{id}/images:
    post:
      summary: Upload an image for a product (admin)
      operationId: adminUploadProductImage
      security: [ { bearerAuth: [] } ]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:     { type: string, format: binary }
                altText:  { type: string }
                position: { type: integer }
              required: [file]
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Image" }
        "400":
          description: Invalid input
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new user.
      operationId: authRegister
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/login:
    post:
      tags: [auth]
      summary: Log in with email + password.
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/verify-email:
    post:
      tags: [auth]
      summary: Consume an email verify token.
      operationId: authVerifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyEmailRequest" }
      responses:
        "200":
          description: OK
        "400": { $ref: "#/components/responses/BadRequest" }
  /auth/verify-email/resend:
    post:
      tags: [auth]
      summary: Re-send the verify email (always returns 202).
      operationId: authResendVerifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResendVerifyRequest" }
      responses:
        "202":
          description: Accepted
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/password-reset/request:
    post:
      tags: [auth]
      summary: Request a password reset email (always returns 202).
      operationId: authPasswordResetRequest
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordResetRequest" }
      responses:
        "202":
          description: Accepted
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/password-reset/confirm:
    post:
      tags: [auth]
      summary: Consume a reset token and set a new password.
      operationId: authPasswordResetConfirm
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordResetConfirm" }
      responses:
        "200":
          description: OK
        "400": { $ref: "#/components/responses/BadRequest" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /auth/csrf:
    get:
      tags: [auth]
      summary: Issue a CSRF cookie + body token (bootstrap before login).
      operationId: authCsrf
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [csrf_token]
                properties:
                  csrf_token: { type: string }
  /auth/logout:
    post:
      tags: [account]
      summary: Logout the current session.
      operationId: authLogout
      security: [{ session: [] }]
      responses:
        "204":
          description: No Content
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /auth/sessions/all:
    delete:
      tags: [account]
      summary: Revoke all sessions for the current user.
      operationId: authRevokeAllSessions
      security: [{ session: [] }]
      responses:
        "204":
          description: No Content
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /me:
    get:
      tags: [account]
      summary: Get the current user.
      operationId: getMe
      security: [{ session: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      tags: [account]
      summary: Update editable fields on the current user.
      operationId: patchMe
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProfileRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /me/change-password:
    post:
      tags: [account]
      summary: Change password (requires current password).
      operationId: changePassword
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordRequest" }
      responses:
        "204":
          description: No Content
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Unprocessable" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    session:
      type: apiKey
      in: cookie
      name: session_id
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unprocessable:
      description: Unprocessable Entity
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    RateLimited:
      description: Too Many Requests
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    Money:
      type: object
      properties:
        amountCents: { type: integer, minimum: 0 }
        currency: { type: string, minLength: 3, maxLength: 3 }
      required: [amountCents, currency]
    Variant:
      type: object
      properties:
        id: { type: string, format: uuid }
        sku: { type: string }
        size: { type: string }
        color: { type: string }
        price: { $ref: "#/components/schemas/Money" }
      required: [id, sku, size, color]
    Image:
      type: object
      properties:
        id: { type: string, format: uuid }
        url: { type: string }
        altText: { type: string }
        position: { type: integer }
        variants: { $ref: "#/components/schemas/ImageVariants" }
      required: [id, url, altText, position]
    ImageVariants:
      type: object
      properties:
        original: { type: string }
        thumb:    { type: string }
        medium:   { type: string }
        large:    { type: string }
      required: [original, thumb, medium, large]
    Product:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        brand: { type: string }
        categoryId: { type: string, format: uuid }
        basePrice: { $ref: "#/components/schemas/Money" }
        status: { type: string, enum: [draft, published, archived] }
        variants:
          type: array
          items: { $ref: "#/components/schemas/Variant" }
        images:
          type: array
          items: { $ref: "#/components/schemas/Image" }
      required: [id, slug, name, categoryId, basePrice, status, variants, images]
    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        parentId: { type: string, format: uuid, nullable: true }
      required: [id, slug, name]
    ProductWriteBody:
      type: object
      properties:
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        brand: { type: string }
        categoryId: { type: string, format: uuid }
        basePriceCents: { type: integer, minimum: 0 }
        currency: { type: string, minLength: 3, maxLength: 3 }
        status: { type: string, enum: [draft, published, archived] }
        variants:
          type: array
          items:
            type: object
            properties:
              sku: { type: string }
              size: { type: string }
              color: { type: string }
              priceCents: { type: integer, minimum: 0 }
            required: [sku, size, color]
        images:
          type: array
          items:
            type: object
            properties:
              url: { type: string }
              altText: { type: string }
              position: { type: integer }
            required: [url]
      required: [slug, name, categoryId, basePriceCents, currency, status]
    CategoryWriteBody:
      type: object
      properties:
        slug: { type: string }
        name: { type: string }
        parentId: { type: string, format: uuid, nullable: true }
      required: [slug, name]
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
          required: [code, message]
      required: [error]
    UserResponse:
      type: object
      required: [id, email, name]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        email_verified_at:
          type: string
          format: date-time
          nullable: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        name: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
        remember: { type: boolean, default: false }
    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token: { type: string }
    ResendVerifyRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    PasswordResetConfirm:
      type: object
      required: [token, new_password]
      properties:
        token: { type: string }
        new_password: { type: string, minLength: 8 }
    UpdateProfileRequest:
      type: object
      properties:
        name: { type: string }
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: { type: string }
        new_password: { type: string, minLength: 8 }